# mbw-app/src/test/resources/openapi/api-docs.snapshot.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition"
  },
  "paths" : {
    "/api/v1/sms-codes" : {
      "post" : {
        "tags" : [ "account-sms-code-controller" ],
        "operationId" : "requestSmsCode",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RequestSmsCodeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/auth/refresh-token" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "refreshToken",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout-all" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "logoutAll",
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/auth/cancel-deletion" : {
      "post" : {
        "tags" : [ "cancel-deletion-controller" ],
        "operationId" : "cancel",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CancelDeletionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/cancel-deletion/sms-codes" : {
      "post" : {
        "tags" : [ "cancel-deletion-controller" ],
        "operationId" : "sendCode",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendCancelDeletionCodeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/accounts/phone-sms-auth" : {
      "post" : {
        "tags" : [ "account-auth-controller" ],
        "operationId" : "phoneSmsAuth",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PhoneSmsAuthRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/me/deletion" : {
      "post" : {
        "tags" : [ "account-deletion-controller" ],
        "operationId" : "delete",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/accounts/me/deletion-codes" : {
      "post" : {
        "tags" : [ "account-deletion-controller" ],
        "operationId" : "sendCode_1",
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/accounts/me" : {
      "get" : {
        "tags" : [ "account-profile-controller" ],
        "operationId" : "getMe",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountProfileResponse"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "account-profile-controller" ],
        "operationId" : "patchMe",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateDisplayNameRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/devices" : {
      "get" : {
        "tags" : [ "device-management-controller" ],
        "operationId" : "list",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeviceListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/devices/{recordId}" : {
      "delete" : {
        "tags" : [ "device-management-controller" ],
        "operationId" : "revoke",
        "parameters" : [ {
          "name" : "recordId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "RequestSmsCodeRequest" : {
        "type" : "object",
        "properties" : {
          "phone" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "phone" ]
      },
      "RefreshTokenRequest" : {
        "type" : "object",
        "properties" : {
          "refreshToken" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "refreshToken" ]
      },
      "LoginResponse" : {
        "type" : "object",
        "properties" : {
          "accountId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "accessToken" : {
            "type" : "string"
          },
          "refreshToken" : {
            "type" : "string"
          }
        }
      },
      "CancelDeletionRequest" : {
        "type" : "object",
        "properties" : {
          "phone" : {
            "type" : "string",
            "minLength" : 1,
            "pattern" : "^\\+\\d{8,15}$"
          },
          "code" : {
            "type" : "string",
            "minLength" : 1,
            "pattern" : "\\d{6}"
          }
        },
        "required" : [ "code", "phone" ]
      },
      "SendCancelDeletionCodeRequest" : {
        "type" : "object",
        "properties" : {
          "phone" : {
            "type" : "string",
            "minLength" : 1,
            "pattern" : "^\\+\\d{8,15}$"
          }
        },
        "required" : [ "phone" ]
      },
      "PhoneSmsAuthRequest" : {
        "type" : "object",
        "properties" : {
          "phone" : {
            "type" : "string",
            "minLength" : 1
          },
          "code" : {
            "type" : "string",
            "minLength" : 1,
            "pattern" : "^\\d{6}$"
          }
        },
        "required" : [ "code", "phone" ]
      },
      "DeleteAccountRequest" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "minLength" : 1,
            "pattern" : "\\d{6}"
          }
        },
        "required" : [ "code" ]
      },
      "UpdateDisplayNameRequest" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "type" : "string",
            "maxLength" : 64,
            "minLength" : 0
          }
        },
        "required" : [ "displayName" ]
      },
      "AccountProfileResponse" : {
        "type" : "object",
        "properties" : {
          "accountId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "phone" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "DeviceItemResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "deviceId" : {
            "type" : "string"
          },
          "deviceName" : {
            "type" : "string"
          },
          "deviceType" : {
            "type" : "string",
            "enum" : [ "PHONE", "TABLET", "DESKTOP", "WEB", "UNKNOWN" ]
          },
          "location" : {
            "type" : "string"
          },
          "loginMethod" : {
            "type" : "string",
            "enum" : [ "PHONE_SMS", "GOOGLE", "APPLE", "WECHAT" ]
          },
          "lastActiveAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isCurrent" : {
            "type" : "boolean"
          }
        }
      },
      "DeviceListResponse" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DeviceItemResponse"
            }
          }
        }
      }
    }
  }
}